Make second parameters truly optional
With the recent updates to this package, I am seeing the following error when using the `hideOthers` function:
`error TS2554: Expected 2-3 arguments, but got 1.`
I believe this second argument is meant to be optional, since it allows `undefined`, but since it doesn't use a `?` or a default value, it is not optional. This issue also applies to `inertOthers` and `suppressOthers`. I think the best solution is just to add `?` and remove `| undefined`. What do you think?